requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope
asttokens
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm
debugpy
decorator
distro==1.9.0
exceptiongroup
executing
fastapi==0.115.12
fastembed==0.6.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
griffe==1.7.2
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
jedi
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client
jupyter_core
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib-inline
mcp==1.6.0
mmh3==5.1.0
mpmath==1.3.0
narwhals==1.35.0
nest_asyncio
numpy==2.2.5
onnxruntime==1.21.1
openai==1.75.0
openai-agents==0.0.11
packaging==24.2
pandas==2.2.3
parso
pexpect
pickleshare
pillow==11.2.1
platformdirs
portalocker==2.10.1
prompt_toolkit
protobuf==5.29.4
psutil
ptyprocess
pure_eval
py_rust_stemmers==0.1.5
pyarrow==19.0.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
Pygments
python-dateutil
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq
qdrant-client==1.13.3
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
setuptools==75.8.0
six
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.2.1
stack_data
starlette==0.46.2
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tornado
tqdm==4.67.1
traitlets
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth
wheel==0.45.1
zipp
